+++ src/steps/__tests__/index.test.ts
@@ -30,7 +30,7 @@ test('steps', async () => {
     name: 'steps',
   });
 
-  const nowTimestamp = 1602457230303; // '2020-10-11T23:00:30.303Z';
+  const nowTimestamp = 1599865230000; // '2020-09-11T23:00:30Z';
   jest.spyOn(Date, 'now').mockImplementation(() => nowTimestamp);
 
   const context = createMockStepExecutionContext<QualysIntegrationConfig>({