+++ src/util/date.ts
@@ -1,7 +1,6 @@
 const MILLISECONDS_ONE_DAY = 1000 * 60 * 60 * 24;
-const MAX_LAST_SCAN_AGE = 30 * MILLISECONDS_ONE_DAY;
 
-function getVmScanSince() {
+function getScannedSinceMs(days: number): number {
   if (process.env.VM_SCAN_SINCE_DAYS) {
     return parseInt(process.env.VM_SCAN_SINCE_DAYS) * MILLISECONDS_ONE_DAY;
   }
@@ -10,11 +9,15 @@ function getVmScanSince() {
     return parseInt(process.env.VM_SCAN_SINCE_MS);
   }
 
-  return MAX_LAST_SCAN_AGE;
+  return days * MILLISECONDS_ONE_DAY;
 }
 
-export function getVmScanSinceDate() {
-  return new Date(Date.now() - getVmScanSince())
+/**
+ * @param scannedSinceDays the number of days since now when the environment
+ * does not specify
+ */
+export function getScannedSinceDate(scannedSinceDays: number) {
+  return new Date(Date.now() - getScannedSinceMs(scannedSinceDays))
     .toISOString()
     .replace(/\.\d{1,3}/, '');
 }