+++ src/validateInvocation.ts
@@ -5,8 +5,10 @@ import {
   IntegrationValidationError,
 } from '@jupiterone/integration-sdk-core';
 
+import { DEFAULT_SCANNED_SINCE_DAYS } from './constants';
 import { createQualysAPIClient } from './provider';
 import { QualysIntegrationConfig } from './types';
+import { getScannedSinceDate } from './util/date';
 
 const REQUIRED_PROPERTIES = [
   'qualysUsername',
@@ -14,6 +16,8 @@ const REQUIRED_PROPERTIES = [
   'qualysApiUrl',
 ];
 
+// TODO: Support numeric values in IntegrationInstanceConfigFieldMap
+// TOOD: Remove config modification code once numeric types supported
 export default async function validateInvocation({
   logger,
   instance,
@@ -35,6 +39,27 @@ export default async function validateInvocation({
       'Invalid API URL: ' + config.qualysApiUrl,
     );
   }
+
+  if (typeof config.minScannedSinceDays === 'string') {
+    const minScannedSinceDays = Number(config.minScannedSinceDays);
+    if (
+      !minScannedSinceDays &&
+      !!(config.minScannedSinceDays as string).trim()
+    ) {
+      throw new IntegrationValidationError(
+        'Invalid minScannedSinceDays: ' + config.minScannedSinceDays,
+      );
+    }
+    config.minScannedSinceDays = minScannedSinceDays;
+  }
+
+  if (!config.minScannedSinceDays) {
+    config.minScannedSinceDays = DEFAULT_SCANNED_SINCE_DAYS;
+  }
+
+  config.minScannedSinceISODate = getScannedSinceDate(
+    config.minScannedSinceDays,
+  );
 
   const client = createQualysAPIClient(logger, instance.config);
   await client.verifyAuthentication();